+++ tests/test_get_dust.py
@@ -1,13 +1,12 @@
 #!/usr/bin/env python3
-import unittest
+import pytest
 import subprocess as sub
 from bin import get_dust
 
 
-class TestMJD(unittest.TestCase):
-    def test_main(self):
-        sub.call(get_dust.__file__, shell=True)
+def test_main():
+    sub.call(get_dust.__file__, shell=True)
 
 
 if __name__ == '__main__':
-    unittest.main()
+    pytest.main()