+++ tests/test_tpm_fetch.py
@@ -1,32 +1,32 @@
 #!/usr/bin/env python3
-import unittest
+import pytest
 from pathlib import Path
 from bin import tpm_fetch
 
 
-class TestTPMFetch(unittest.TestCase):
-    def setUp(self):
-        self.file = Path(__file__).absolute().parent.parent / 'bin/tpm_fetch.py'
+class DummyArgs(object):
+    pass
 
-        class DummyArgs(object):
-            pass
 
-        self.args = DummyArgs()
-        self.args.t1 = '2020-11-05T06:00:00'
-        self.args.t2 = '2020-11-06T06:30:00'
-        self.args.mjd = None
-        self.args.channels = ['alt_pos', 'az_pos']
-        self.args.verbose = True
+args = DummyArgs()
+args.t1 = '2020-11-05T06:00:00'
+args.t2 = '2020-11-06T06:30:00'
+args.mjd = None
+args.channels = ['alt_pos', 'az_pos']
+args.verbose = True
+
+
+class TestTPMFetch():
 
     def test_time_range(self):
-        tpm_fetch.main(args=self.args)
+        tpm_fetch.main(args=args)
 
     def test_one_mjd(self):
-        self.args.t1 = None
-        self.args.t2 = None
-        self.args.mjd = 59159
-        tpm_fetch.main(args=self.args)
+        args.t1 = None
+        args.t2 = None
+        args.mjd = 59159
+        tpm_fetch.main(args=args)
 
 
 if __name__ == '__main__':
-    unittest.main()
+    pytest.main()