+++ tests/test_wavemid.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
-import unittest
+import pytest
 import subprocess as sub
 from pathlib import Path
 
 
-class TestWAVEMID(unittest.TestCase):
+class TestWAVEMID():
 
     def test_4cam(self):
         """Runs a test from a normal 4-camera xmid
@@ -14,7 +14,7 @@ class TestWAVEMID(unittest.TestCase):
         x = sub.check_output(' '.join([(bin_dir / 'WAVEMID').as_posix(),
                                        '4907.9', '7989.9', '5000.8', '8037.1']),
                              shell=True)
-        self.assertNotIn(x.decode('utf-8'), 'Traceback')
+        assert "Traceback" not in x.decode('utf-8')
 
     def test_2cam(self):
         """Runs a test from a normal 4-camera xmid
@@ -23,8 +23,8 @@ class TestWAVEMID(unittest.TestCase):
         bin_dir = here.parent / 'bin'
         x = sub.check_output([(bin_dir / 'WAVEMID').as_posix(), '4907.9',
                               '7989.9'])
-        self.assertNotIn(x.decode('utf-8'), 'Traceback')
+        assert "Traceback" not in x.decode('utf-8')
 
 
 if __name__ == '__main__':
-    unittest.main()
+    pytest.main()