+++ tests/test_xmid.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
-import unittest
+import pytest
 import subprocess as sub
 from pathlib import Path
 
 
-class TestXMID(unittest.TestCase):
+class TestXMID():
 
     def test_4cam(self):
         """Runs a test from a normal 4-camera xmid
@@ -13,7 +13,7 @@ class TestXMID(unittest.TestCase):
         bin_dir = here.parent / 'bin'
         x = sub.check_output([bin_dir / 'XMID', '2043.7', '2046.8', '2042.7',
                               '2019.0'])
-        self.assertNotIn(x.decode('utf-8'), 'Traceback')
+        assert "Traceback" not in x.decode('utf-8')
 
     def test_2cam(self):
         """Runs a test from a normal 4-camera xmid
@@ -21,8 +21,8 @@ class TestXMID(unittest.TestCase):
         here = Path(__file__).parent.absolute()
         bin_dir = here.parent / 'bin'
         x = sub.check_output([bin_dir / 'XMID', '2049.2', '2046.9'])
-        self.assertNotIn(x.decode('utf-8'), 'Traceback')
+        assert "Traceback" not in x.decode('utf-8')
 
 
 if __name__ == '__main__':
-    unittest.main()
+    pytest.main()