+++ CHANGELOG.md
@@ -529,3 +529,9 @@
 ## [3.7.14] - 2022-04-19
 
 - Whoops
+
+## [3.8.0] - 2022-05-31
+
+- Switched to pytest
+- Fixed to log support on problematic nights
+- moved dust.flux to actors bucket